:recycle: Refactor: 'authApi.js'를 api 경로로 이동, 'tokenStorage.js'를 stores 파일 경로로 이동하도록 변경
